+++ apps/react-web/src/locales/zh_CN.ts
@@ -1,5 +1,0 @@
-const locale = {
-  foo: 'foo',
-}
-
-export default locale